Obtention des representatives docs

+++ util/topicModeling/getTopic.py
@@ -1,13 +1,16 @@
 from bertopic import BERTopic
-from sklearn.datasets import fetch_20newsgroups
 import os
 from PyPDF2 import PdfReader
 import pickle
 import pandas as pd
 from sklearn.feature_extraction.text import CountVectorizer
 #os.environ["TOKENIZERS_PARALLELISM"] = "false"
+import plotly.io as pio
+import csv
+import sys
 
-pdf_folder = '/media/gad/D231-2070/PDF'
+
+pdf_folder = '/home/gad/Documents/PDF'
 
 def docsMaker(pdf_folder):
     docs=[]
@@ -26,6 +29,51 @@ def docsMaker(pdf_folder):
     with open("docs_dl.txt", "wb") as fp:   #Pickling
         pickle.dump(docs, fp)
     return docs
+
+def hash_maker():
+    hashmap={}
+    pdfs=os.listdir(pdf_folder)
+    with open("docs_dl.txt", "rb") as fp:   # Unpickling
+        docs_dl = pickle.load(fp)
+    fichier=0
+    for doc in docs_dl:
+        hashmap[doc]=pdfs[fichier]
+        fichier +=1
+
+    return hashmap
+        
+def get_representative_docs(representative_docs,hashmap):
+    docs=[]
+    for doc in representative_docs:
+        docs.append(hashmap[doc])
+    return docs
+
+def representative_docs_gen(topic_model):
+    csv.field_size_limit(sys.maxsize)
+    hashmap=hash_maker()
+    representative_docs=csv_gen(topic_model)
+    docs=[]
+    for representative in representative_docs:
+        docs.append(get_representative_docs(representative,hashmap))
+    
+    input_file = 'BERTopic_output.csv'
+    output_file = 'BERTopic_modified.csv'
+
+    with open(input_file, mode='r', newline='') as infile, open(output_file, mode='w', newline='') as outfile:
+        reader = csv.reader(infile)
+        writer = csv.writer(outfile)
+        next(reader,None)
+        for i, row in enumerate(reader):
+            if i < len(docs):
+                if len(row) > 4:  
+                    row[4] = docs[i]
+                else:  # Ajoute des colonnes vides si nécessaire pour atteindre la 5ème colonne
+                    row.extend([''] * (5 - len(row)))
+                    row[4] = docs[i]
+            writer.writerow(row)
+
+    print(f"Le fichier CSV a été mis à jour et enregistré sous le nom {output_file}.")
+
 
 def model_train():
     with open("docs_dl.txt", "rb") as fp:   # Unpickling
@@ -48,7 +96,11 @@ def model_train():
 
 def csv_gen(topic_model):
     df=topic_model.get_topic_info()
-    df.to_csv('BERTopic_output.csv',index=False)
+    if not os.path.exists("BERTopic_output.csv"):
+        df.to_csv('BERTopic_output.csv',index=False)
+    representative_docs= df["Representative_Docs"]
+    return representative_docs
+
 
 
 def gen_heatmap(topic_model):
@@ -59,12 +111,14 @@ def gen_visualize_documents(topic_model,docs_dl):
     fig=topic_model.visualize_documents(docs_dl,
                                         topics=list(range(30)),
                                         height=600)
-    fig.write_image("visualize_topic/visualize_docs.png")
+    pio.write_image(fig, "visualize_topic/output.png", engine="kaleido")
 
 
 
 if __name__=="__main__":
     topic_model,docs_dl=model_train()
+    representative_docs_gen(topic_model)
+    
     #gen_heatmap(topic_model)
-    gen_visualize_documents(topic_model,docs_dl)
+    #gen_visualize_documents(topic_model,docs_dl)
